give weekday generator its own controller
