+++ app/views/assignments/weekday_generators/prompt.html.haml
@@ -2,7 +2,7 @@
   %h1 Generate by Weekday
   %hr
   = form_with model: @generator,
-              url: generate_by_weekday_roster_assignments_path(@roster),
+              url: roster_assignments_generate_by_weekday_path(@roster),
               local: true do |f|
     .mb-3
       = f.label :user_id, class: 'form-label'
