+++ app/views/layouts/_nav.html.haml
@@ -18,7 +18,7 @@
                         'data-bs-toggle': 'dropdown', 'aria-expanded': 'false'
               .dropdown-menu{ 'aria-labelledby': 'navbar-generate-nav-link' }
                 = link_to 'Rotation', rotation_generator_roster_assignments_path(@roster), class: 'dropdown-item'
-                = link_to 'By Weekday', generate_by_weekday_roster_assignments_path(@roster), class: 'dropdown-item'
+                = link_to 'By Weekday', roster_assignments_generate_by_weekday_path(@roster), class: 'dropdown-item'
       %ul.navbar-nav
         - if Current.user.present?
           = nav_link_item Current.user.full_name,
